+++ linear_congruential.h
@@ -4,15 +4,15 @@ class LinearCongruentialGenerator {
 	long long modulus;
 	int multiplier;
 	int increment;
-	int state {0}; // default seed is 0
+	long long state {0}; // default seed is 0
 
 	void updateState(); // computes the next state; called by randint and randfloat
 
 public:
 	LinearCongruentialGenerator(long long modulus, int multiplier, int increment);
-	void seed(int seed) { state = seed; };
+	void seed(long long seed) { state = seed; };
 	long long getModulus() { return modulus; };
 
-	int rand();
+	long long rand();
 	float randfloat(); // return a number between 0 and 1
 };
